measure/control the video projector via homematic
The Optoma HD25-LV cannot be reliably controlled via its serial
interface. After a couple of days, it will just not respond to any
commands anymore and needs a hard-reset to continue working.

Hence, I bought a homematic HM-ES-PMSw1-Pl and attached the video
projector to it. The HM-ES-PMSw1-Pl measures power and can control the
power state of the video projector. This way, I can turn on the video
projector only when necessary, avoiding serial interface troubles.